Rename packages to follow Spring Boot conventions
