+++ spring-boot/legacy-starter/src/main/java/org/togglz/spring/boot/legacy/actuate/autoconfigure/TogglzEndpointAutoConfiguration.java
@@ -14,34 +14,35 @@
  * limitations under the License.
  */
 
-package org.togglz.spring.boot.autoconfigure.actuator;
+package org.togglz.spring.boot.legacy.actuate.autoconfigure;
 
-import org.springframework.boot.actuate.autoconfigure.endpoint.condition.ConditionalOnEnabledEndpoint;
-import org.springframework.boot.actuate.endpoint.annotation.Endpoint;
+import org.springframework.boot.actuate.endpoint.AbstractEndpoint;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.togglz.core.manager.FeatureManager;
 import org.togglz.spring.boot.autoconfigure.TogglzAutoConfiguration;
+import org.togglz.spring.boot.legacy.actuate.TogglzEndpoint;
 
 /**
- * {@link EnableAutoConfiguration Auto-configuration} for Togglz Endpoint (Spring Boot 2.x).
+ * {@link EnableAutoConfiguration Auto-configuration} for Togglz Endpoint (Spring Boot 1.5).
  *
  * @author Rui Figueira
  */
 @Configuration
-@ConditionalOnClass(Endpoint.class)
+@ConditionalOnClass(AbstractEndpoint.class)
 @AutoConfigureAfter(TogglzAutoConfiguration.class)
 public class TogglzEndpointAutoConfiguration {
 
     @Bean
     @ConditionalOnBean(FeatureManager.class)
     @ConditionalOnMissingBean
-    @ConditionalOnEnabledEndpoint
+    @ConditionalOnProperty(prefix = "togglz.endpoint", name = "enabled", matchIfMissing = true)
     public TogglzEndpoint togglzEndpoint(FeatureManager featureManager) {
         return new TogglzEndpoint(featureManager);
     }
