+++ spring-boot/legacy-starter/src/test/java/org/togglz/spring/boot/legacy/actuate/BaseTest.java
@@ -14,15 +14,18 @@
  * limitations under the License.
  */
 
-package org.togglz.spring.boot.autoconfigure;
+package org.togglz.spring.boot.legacy.actuate;
 
-import org.springframework.boot.actuate.autoconfigure.endpoint.EndpointAutoConfiguration;
-import org.springframework.boot.actuate.autoconfigure.endpoint.web.WebEndpointAutoConfiguration;
-import org.springframework.boot.actuate.autoconfigure.web.server.ManagementContextAutoConfiguration;
-import org.springframework.boot.autoconfigure.AutoConfigurations;
-import org.springframework.boot.test.context.runner.WebApplicationContextRunner;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+import org.junit.After;
+import org.springframework.boot.actuate.autoconfigure.ManagementServerPropertiesAutoConfiguration;
+import org.springframework.boot.test.util.EnvironmentTestUtils;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
 import org.togglz.core.Feature;
 import org.togglz.core.activation.Parameter;
 import org.togglz.core.manager.EnumBasedFeatureProvider;
@@ -31,20 +34,40 @@ import org.togglz.core.spi.ActivationStrategy;
 import org.togglz.core.spi.FeatureProvider;
 import org.togglz.core.user.FeatureUser;
 import org.togglz.spring.boot.autoconfigure.TogglzAutoConfiguration;
-import org.togglz.spring.boot.autoconfigure.actuator.TogglzEndpointAutoConfiguration;
-import org.togglz.spring.boot.autoconfigure.management.TogglzManagementContextConfiguration;
+import org.togglz.spring.boot.legacy.actuate.autoconfigure.TogglzEndpointAutoConfiguration;
+import org.togglz.spring.boot.legacy.actuate.autoconfigure.TogglzManagementContextConfiguration;
 
 public class BaseTest {
 
-    protected final WebApplicationContextRunner contextRunner = new WebApplicationContextRunner()
-            .withConfiguration(AutoConfigurations.of(
-                    ManagementContextAutoConfiguration.class,
-                    EndpointAutoConfiguration.class,
-                    WebEndpointAutoConfiguration.class,
-                    TogglzAutoConfiguration.class,
-                    TogglzEndpointAutoConfiguration.class,
-                    TogglzManagementContextConfiguration.class
-                    ));
+    protected AnnotationConfigWebApplicationContext context;
+
+    @After
+    public void tearDown() {
+        if (this.context != null) {
+            this.context.close();
+        }
+    }
+
+    protected void loadWithDefaults(String... environment) {
+        loadWithDefaults(new Class[]{}, environment);
+    }
+
+    protected void loadWithDefaults(Class<?>[] configs, String... environment) {
+        List<Class<?>> mergedConfigs = new ArrayList<>();
+        mergedConfigs.add(ManagementServerPropertiesAutoConfiguration.class);
+        mergedConfigs.add(TogglzAutoConfiguration.class);
+        mergedConfigs.add(TogglzEndpointAutoConfiguration.class);
+        mergedConfigs.add(TogglzManagementContextConfiguration.class);
+        mergedConfigs.addAll(Arrays.asList(configs));
+        load(mergedConfigs.toArray(new Class[mergedConfigs.size()]), environment);
+    }
+
+    protected void load(Class<?>[] configs, String... environment) {
+        this.context = new AnnotationConfigWebApplicationContext();
+        this.context.register(configs);
+        EnvironmentTestUtils.addEnvironment(this.context, environment);
+        this.context.refresh();
+    }
 
     protected enum MyFeatures implements Feature {
         FEATURE_ONE,
@@ -91,14 +114,6 @@ public class BaseTest {
         public CustomActivationStrategy customActivationStrategy() {
             return new CustomActivationStrategy();
         }
-    }
-
-    protected WebApplicationContextRunner contextRunnerWithFeatureProviderConfig() {
-        return contextRunner
-            .withPropertyValues(
-                    "togglz.console.enabled: true",
-                    "togglz.console.use-management-port: true")
-            .withUserConfiguration(FeatureProviderConfig.class);
     }
 
 }
