+++ spring-boot/legacy-starter/src/test/java/org/togglz/spring/boot/legacy/actuate/autoconfigure/TogglzManagementContextConfigurationTest.java
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package org.togglz.legacy.spring.boot.autoconfigure.management;
+package org.togglz.spring.boot.legacy.actuate.autoconfigure;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 import org.springframework.boot.web.servlet.ServletRegistrationBean;
-import org.togglz.legacy.spring.boot.autoconfigure.BaseTest;
 import org.togglz.spring.boot.autoconfigure.TogglzAutoConfiguration;
+import org.togglz.spring.boot.legacy.actuate.BaseTest;
 
 /**
  * Tests for {@link TogglzAutoConfiguration}.
