+++ spring-boot/starter/src/main/java/org/togglz/spring/boot/actuate/TogglzEndpoint.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.togglz.spring.boot.autoconfigure.actuator;
+package org.togglz.spring.boot.actuate;
 
 import java.util.ArrayList;
 import java.util.Collections;
