+++ spring-boot/starter/src/main/resources/META-INF/spring.factories
@@ -1,6 +1,6 @@
 # Auto Configure
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.togglz.spring.boot.autoconfigure.actuator.TogglzEndpointAutoConfiguration
+org.togglz.spring.boot.actuate.autoconfigure.TogglzEndpointAutoConfiguration
 
 org.springframework.boot.actuate.autoconfigure.web.ManagementContextConfiguration=\
-org.togglz.spring.boot.autoconfigure.management.TogglzManagementContextConfiguration
+org.togglz.spring.boot.actuate.autoconfigure.TogglzManagementContextConfiguration
